Integrated raw_store into store
